% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.discretisation}
\alias{.discretisation}
\title{Discretize Continuous Eigenvectors in Spectral Clustering}
\usage{
.discretisation(eigenVectors)
}
\arguments{
\item{eigenVectors}{A matrix of continuous eigenvectors.}
}
\value{
A list containing:
\item{discrete}{Discrete cluster indicator matrix}
\item{continuous}{Row-normalized eigenvectors}
}
\description{
Internal function for converting continuous eigenvectors to discrete
cluster indicators using Yu & Shi (2003) algorithm.
}
\keyword{internal}
