% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.mutualInformation}
\alias{.mutualInformation}
\title{Compute Mutual Information Between Two Label Vectors}
\usage{
.mutualInformation(x, y)
}
\arguments{
\item{x}{A vector of discrete labels.}

\item{y}{A second vector of discrete labels.}
}
\value{
Numeric value of mutual information (bits).
}
\description{
Internal function for calculating mutual information.
}
\keyword{internal}
