% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.repmat}
\alias{.repmat}
\title{Replicate Matrix (MATLAB \code{repmat} Equivalent)}
\usage{
.repmat(X, m, n)
}
\arguments{
\item{X}{A scalar, vector, or matrix to be replicated.}

\item{m}{Number of row replications.}

\item{n}{Number of column replications.}
}
\value{
Replicated matrix.
}
\description{
Internal helper function for matrix replication.
}
\keyword{internal}
