% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{entropy}
\alias{entropy}
\title{Compute Shannon Entropy of a Label Vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{A vector of categorical labels.}
}
\value{
Shannon entropy (base-2).
}
\description{
Compute Shannon Entropy of a Label Vector
}
