% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KM.R
\name{get_color}
\alias{get_color}
\title{Generate Colors from Predefined Palettes}
\usage{
get_color(palette, n = 12)
}
\arguments{
\item{palette}{A palette name such as "jama_classic", "nature",
"jco", "lancet", etc., or a custom vector of colors.}

\item{n}{Number of colors to return.}
}
\value{
A character vector of colors.
}
\description{
Returns a vector of colors generated from a predefined palette.
If \code{palette} is already a vector of colors, it is returned directly.
}
