% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{perc}
\alias{perc}
\title{Convert Values to Proportions}
\usage{
perc(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector where each element is divided by sum(x).
}
\description{
Convert Values to Proportions
}
