% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_diffusion.R
\name{self.diffusion}
\alias{self.diffusion}
\title{Self-Diffusion on Local Scaling Affinity}
\usage{
self.diffusion(A, K)
}
\arguments{
\item{A}{A square affinity matrix.}

\item{K}{Integer. Number of iteration steps used to determine the dominate set.}

\item{aff.matrix}{A square affinity matrix.}

\item{NR.OF.KNN}{Integer. Number of nearest neighbors to retain in each row.}

\item{W}{A square matrix.}

\item{w}{A square symmetric matrix.}

\item{type}{Character string. The normalization type:
\itemize{
\item \code{"ave"}: return \eqn{D^{-1} W}
\item \code{"gph"}: return \eqn{D^{-1/2} W D^{-1/2}}
}}
}
\value{
A diffused affinity matrix.

A symmetric dominate-set matrix with top-K values preserved.

A transformed matrix representing transition fields.

A normalized kernel matrix.
}
\description{
This function performs self-diffusion on a given affinity matrix using
a diffusion process controlled by eigen-decomposition.
}
\examples{
data(GSE62254_data)
data1 <- apply(GSE62254_gene.expression, 2, function(x) (x-mean(x))/sd(x))
A <- affs(data1)
self.diffusion(A, 4)
Compute Dominate Set for Affinity Matrix

This function computes the dominate set matrix for a given affinity matrix,
based on selecting the top-K nearest neighbors for each row.

}
\keyword{affinity}
\keyword{diffusion}
\keyword{dominate-set}
\keyword{graph}
\keyword{kernel}
\keyword{knn}
\keyword{normalization}
\keyword{transition-field}
